requirements.txt
python-telegram-bot
google-genai
python-dotenv
requests
beautifulsoup4
playwright
docker
gitpython
